Check if data exists before executing commands that would fail otherwise

+++ resources/lib/utils.py
@@ -85,14 +85,17 @@ def remove_cached_art(art):
         #Get cached image name to remove                  
         db.execute("SELECT cachedurl FROM texture WHERE url = '%s';" %art)
         data = db.fetchone()
-        #log(4, 'Removing Cached Art :: SQL Output: %s' %data[0]) -> Causes an error on first time configuring
         
-        file_to_delete = os.path.join(_tbn_path, data[0])
-        log(4, 'Removing Cached Art :: File to be removed: %s' %file_to_delete)
+        if data[0]:
+            
+            log(4, 'Removing Cached Art :: SQL Output: %s' %data[0]) -> Causes an error on first time configuring
+            
+            file_to_delete = os.path.join(_tbn_path, data[0])
+            log(4, 'Removing Cached Art :: File to be removed: %s' %file_to_delete)
 
-        xbmcvfs.delete(file_to_delete)
-        db.execute("DELETE FROM texture WHERE url = '%s';" %art)
-    
+            xbmcvfs.delete(file_to_delete)
+            db.execute("DELETE FROM texture WHERE url = '%s';" %art)
+        
     except lite.Error, e:
         print "Error %s:" % e.args[0]
         #sys.exit(1)
